+++ HISTORY.md
@@ -1,9 +1,13 @@
 unreleased
 ==========
 
+- update: `dnsPrefetchControl` middleware is now default
+
 - fix: make `hpkp` lowercase in documentation
 - fix: update `hpkp` spec URL in readmes
 - fix: update `frameguard` header name in readme
+
+- remove: module aliases. There is now just one way to include each middleware
 
 1.3.0 / 2016-03-01
 ==================